paginate search by id and search by text
